+++ src/spring/HISTORY.md
@@ -1,5 +1,11 @@
 Release History
 ===============
+1.1.11
+---
+* Add command `az spring app deployment enable-remote-debugging`.
+* Add command `az spring app deployment disable-remote-debugging`.
+* Add command `az spring app deployment get-remote-debugging-config`.
+
 1.1.10
 ---
 * Remove `Preview` tag for user-assigned identities of apps.